Separated out the length batch and time batch code from the topK classes to an abstract class and set whether it should be topK or bottomK in the child classes extending them
